Rework the server package structure to make it easier to include as a library
