+++ packages/server/tsconfig.json
@@ -14,7 +14,7 @@
     { "path": "../crypto/tsconfig.build.json" },
     { "path": "../lexicon/tsconfig.build.json" },
     { "path": "../repo/tsconfig.build.json" },
-    { "path": "../xrpc-cli/tsconfig.build.json" },
+    { "path": "../lex-cli/tsconfig.build.json" },
     { "path": "../xrpc-server/tsconfig.build.json" }
   ]
 }
